moved thunks to fetch.js
