+++ code/src/components/Labyrinth.js
@@ -4,7 +4,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { useSelector, useDispatch } from 'react-redux'
 
-import { fetchNext } from 'reducers/game'
+import { fetchNext } from 'reducers/fetch'
 import { Button } from 'lib/Button'
 import { Card } from 'lib/Card'
