+++ code/src/reducers/fetch.js
@@ -1,0 +1,45 @@
+/* eslint-disable object-shorthand */
+/* eslint-disable object-curly-newline */
+import { game } from './game'
+import { ui } from './ui'
+
+export const fetchStart = (User) => {
+  const START_API = 'https://wk16-backend.herokuapp.com/start/'
+
+  return (dispatch) => {
+    dispatch(ui.actions.setLoading(true))
+    fetch(START_API, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ username: User })
+    })
+      .then((res) => res.json())
+      .then((json) => {
+        dispatch(game.actions.setCurrentStep(json))
+        dispatch(ui.actions.setLoading(false))
+        console.log(json)
+      })
+  }
+}
+
+export const fetchNext = (User, direction) => {
+  const NEXT_API = 'https://wk16-backend.herokuapp.com/action/'
+
+  return (dispatch) => {
+    dispatch(ui.actions.setLoading(true))
+    fetch(NEXT_API, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({
+        username: User,
+        type: 'move',
+        direction: direction })
+    })
+      .then((res) => res.json())
+      .then((json) => {
+        dispatch(game.actions.setCurrentStep(json))
+        dispatch(ui.actions.setLoading(false))
+        console.log(json)
+      })
+  }
+}
